requirements.txt
# Astro Runtime includes the following pre-installed providers packages: https://docs.astronomer.io/astro/runtime-image-architecture#provider-packages
beautifulsoup4==4.12.3
selenium==4.18.1
trio==0.25.0
pymongo==4.6.2
apache-airflow-providers-mongo==4.0.0
py-trans==0.6.1
aiofiles==23.2.1
